test: move page tests to routes/__tests__ dir and remove react-router
